Change TF2Class from module to a class
TF2ClassFactory was calling TF2Class.new for all the classes that don't have
class specific statistics.

As TF2Class was a module, it did not have "new" defined on it.

Changed it to a class which all the other classes derive from.

This change in the module usage might only be from 1.9.2, not sure